fix(lyrics): trim_outlier_indices uses max-gap-anywhere, not last-gap
Wall-verify on id=21 4:05 found "in the house of my Father" displayed
for 33.5 s (matched span 230214-263774 with a 15 s held vowel between
"house" and "of" in the middle of the matched range).

The 2026-05-07 trim fix only checked the LAST consecutive gap
(prev→last) when deciding whether to pop the trailing outlier. If the
big gap is in the MIDDLE of the matched range (matched [in, the,
house, of, my, Father] with house→of = 15 s but my→Father = 100 ms),
last-pair-gap is small and the fix incorrectly preserves the entire
range.

New rule: pop trailing while span > LONG_LINE_CAP_MS AND there exists
ANY consecutive gap ≥ TRIM_GAP_MS (3 s). If max-gap-anywhere is below
the threshold, the matches are tight (held notes) and we keep the
whole range — Phase 4 sub-line LCS splits it. If max-gap is above
threshold, an outlier exists somewhere — pop trailing until span fits
the cap or only 1 entry left.

Held-note test from 2026-05-07 (max gap < 3 s) still preserves all
matches. New regression test for 2026-05-08 covers the middle-gap case
(matched [a..f] with the d-gap = 15 s between c and d): pop trailing
[d, e, f] until span fits cap.

Extracts the trim tests to a new sibling file
text_reference_merge_trim_tests.rs so the parent tests file stays
under the 1000-line cap. Trims a few inline doc comments in
text_reference_merge.rs to keep the parent file under cap too.

Pause/resume on live playlist (ytlive) is a separate issue — engine
treats Play/Pause API as OBS scene events; for a custom playlist
SceneOn triggers SelectAndPlay (next video), not resume. Tracked
separately.
