+++ crates/sp-server/src/lyrics/text_reference_merge_tests.rs
@@ -432,105 +432,6 @@ fn absorb_prefix_walks_scan_to_zero_without_match() {
     assert_eq!(emits[0].asr_word_indices, vec![3]);
 }
 
-// ── Phase 2.5: trim_outlier_indices ───────────────────────────────────────────
-
-#[test]
-fn trim_outlier_indices_keeps_tight_match_intact() {
-    let asr_track = asr(vec![
-        make_word("a", 0, 100),
-        make_word("b", 200, 400),
-        make_word("c", 500, 700),
-        make_word("d", 800, 1000),
-    ]);
-    let asr_words = flatten_asr(&asr_track);
-    let mut indices = vec![0, 1, 2, 3];
-    trim_outlier_indices(&mut indices, &asr_words);
-    assert_eq!(indices, vec![0, 1, 2, 3]);
-}
-
-#[test]
-fn trim_outlier_indices_drops_trailing_outlier_past_cap() {
-    // 5 words: [0..4 contiguous within 7.5s] + [5 jumped to 20s].
-    // Span 20s > LONG_LINE_CAP_MS=8s, drop trailing.
-    let asr_track = asr(vec![
-        make_word("a", 0, 500),
-        make_word("b", 1000, 1500),
-        make_word("c", 3000, 3500),
-        make_word("d", 5000, 5500),
-        make_word("e", 7000, 7500),
-        make_word("outlier", 19000, 20000),
-    ]);
-    let asr_words = flatten_asr(&asr_track);
-    let mut indices = vec![0, 1, 2, 3, 4, 5];
-    trim_outlier_indices(&mut indices, &asr_words);
-    // After trim: [0..4] span 7.5s within cap.
-    assert_eq!(indices, vec![0, 1, 2, 3, 4]);
-}
-
-#[test]
-fn trim_outlier_indices_drops_to_single_when_two_entry_span_exceeds_cap() {
-    // 2-entry emit with span > cap — trim drops the trailing outlier so
-    // only [0] remains. Phase 5 will then reject the single-word residual
-    // via its MIN_LINE_DURATION_MS micro-window drop. Reproduces id=132
-    // 2026-05-04 case where Phase 2 LCS for "Holy forever" picked the
-    // first "holy" + a far-later "forever" across multiple sung phrases.
-    let asr_track = asr(vec![
-        make_word("a", 0, 100),
-        make_word("b", 50000, 50100), // 50s gap — exceeds cap
-    ]);
-    let asr_words = flatten_asr(&asr_track);
-    let mut indices = vec![0, 1];
-    trim_outlier_indices(&mut indices, &asr_words);
-    assert_eq!(indices, vec![0]);
-}
-
-#[test]
-fn trim_outlier_indices_keeps_single_entry_intact() {
-    // Single-entry input is untouched (no second word to compare span).
-    let asr_track = asr(vec![make_word("a", 1000, 1500)]);
-    let asr_words = flatten_asr(&asr_track);
-    let mut indices = vec![0];
-    trim_outlier_indices(&mut indices, &asr_words);
-    assert_eq!(indices, vec![0]);
-}
-
-#[test]
-fn trim_outlier_indices_keeps_contiguous_held_notes_past_cap() {
-    // id=21 "Good Shepherd" 2026-05-07: span 12.4 s due to held vowels on
-    // "enemies" (1.24 s) + "forgiveness" (2.40 s), all gaps < 3 s.
-    // Old impl dropped the tail; new gap-guarded impl preserves it.
-    let asr_track = asr(vec![
-        make_word("a", 0, 100),
-        make_word("b", 1000, 1500),
-        make_word("c", 2500, 3000),
-        make_word("d", 4000, 4500),
-        make_word("e", 5500, 6000),
-        make_word("f", 7000, 8000),
-        make_word("g", 8500, 12500), // 4.5 s held note
-    ]);
-    let asr_words = flatten_asr(&asr_track);
-    let mut indices = (0..7).collect::<Vec<_>>();
-    trim_outlier_indices(&mut indices, &asr_words);
-    // All 7 preserved — span 12.5 s > 8 s cap but every gap < 3 s.
-    assert_eq!(indices, (0..7).collect::<Vec<_>>());
-}
-
-#[test]
-fn trim_outlier_indices_handles_unsorted_input() {
-    // Indices arrive ascending after Phase 1 sort, but defensive: trim should
-    // sort before measuring span.
-    let asr_track = asr(vec![
-        make_word("a", 0, 100),
-        make_word("b", 200, 400),
-        make_word("c", 500, 700),
-        make_word("outlier", 20000, 21000),
-    ]);
-    let asr_words = flatten_asr(&asr_track);
-    let mut indices = vec![3, 0, 1, 2];
-    trim_outlier_indices(&mut indices, &asr_words);
-    assert_eq!(indices, vec![0, 1, 2]);
-}
-
 // ── Phase 4: aligned_lines_for_emit ───────────────────────────────────────────
 
 #[test]
